Adding jsvalidator to the repo.
Getting an error http://prntscr.com/pl5mv8 on an app the does not require jsvalidator and is only pulling in the schema loader. Either way the container is looking for jsvaliator which is not a package that is available to it, and must be installed on an application to bypass this error.